requirements.txt
# ARMGUARD RDS V1 — pinned production dependencies
# Generated from active venv on 2026-03-09 (Django 6.0.3, Python 3.x)
# Update with: pip freeze > requirements.txt
# ── Core framework ─────────────────────────────────────────────
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3

# ── Image / QR / PDF ───────────────────────────────────────────
pillow==12.1.1
PyMuPDF==1.27.1     # fitz: used by form_filler.py for PDF template filling
qrcode==8.2
# reportlab removed — pdf_filler1.py was deleted in Session 7; form_filler.py
# uses PyMuPDF (fitz) exclusively. Remove reportlab from venv with:
#   pip uninstall reportlab

# ── Static file serving ────────────────────────────────────────
whitenoise==6.12.0           # L6: serve compressed static files from WSGI process

# ── Environment / utilities ────────────────────────────────────
python-dotenv==1.2.2
# L1 FIX: colorama removed — unused; no code imports it.
charset-normalizer==3.4.5

# ── REST API (G12) ────────────────────────────────────────────────────────────
djangorestframework==3.16.0

# ── MFA / TOTP (G15) ──────────────────────────────────────────────────────────
django-otp==1.7.0

# ── Production WSGI server ────────────────────────────────────────────────────
gunicorn==22.0.0             # WSGI server for production (systemd + Nginx)

# ── Optional: error tracking (uncomment to enable) ─────────────
# sentry-sdk==2.x.x         # L5: uncomment + set SENTRY_DSN env var to enable

# ── PostgreSQL driver (required for production) ────────────────
psycopg2-binary>=2.9.9   # L2 FIX: declared (uncommented); required for PostgreSQL production DB
